ARCH-2/ARCH-3: lint subsystem into hyalo-mdlint; unified MutationJournal for index maintenance (iter-226)
ARCH-2 — lint crate boundary:
- move the five profile linters (changelog/madr/okf/skills/github) plus
  heading_grammar, link_lint (HYALO006) and section_scanner into
  hyalo-mdlint::profiles
- move the schema-validation core of commands/lint.rs (types, violation
  kinds, lint_file/lint_file_with_fix, auto-fix, validate_constraint
  family) into hyalo-mdlint::schema, exposing an in-process lint API;
  commands::lint re-exports for unchanged call sites; output byte-identical

ARCH-3 — index refresh as a property of the write path:
- new commands::journal::MutationJournal replaces save_index_if_dirty
  (8 call sites), tasks.rs::patch_index, and patch_index_for_modified_files;
  every mutating command (set/remove/append/new/mv/task/properties/tags/
  links/lint --fix) records through it, flushed once per invocation,
  always refreshing entry AND persisted link graph
- fixes two latent stale-link-graph bugs: properties rename and tags
  rename under --index never refreshed the link graph
- xtask check-mutation-journal gate: fails CI when a mutating module
  bypasses the journal or a forbidden persistence token reappears
- e2e tests/e2e/index_journal.rs pins the index.rs:439 stale-graph
  regression class via each mutation path

Docs: DEC-226 decision-log entry, CHANGELOG entry, iteration 226 tasks
ticked. Quality gates: fmt, clippy -D warnings, workspace tests,
all xtask check-* gates green.
