+++ crates/hyalo-cli/src/commands/mod.rs
@@ -2,36 +2,29 @@
 pub mod append;
 pub mod backlinks;
 pub mod changelog;
-pub mod changelog_lint;
 pub(crate) mod config;
 pub mod create_index;
 pub mod drop_index;
 pub mod files_from;
 pub mod find;
-pub mod heading_grammar;
 pub mod init;
 pub(crate) mod inputs;
 pub(crate) mod iteration;
-pub mod link_lint;
+pub mod journal;
 pub mod links;
 pub mod lint;
-pub mod lint_github;
 pub mod lint_rules;
 pub mod madr;
-pub mod madr_lint;
 pub(crate) mod managed_region;
 pub(crate) mod mutation;
 pub mod mv;
 pub mod new;
 pub mod okf;
-pub mod okf_lint;
 pub mod profiles;
 pub mod properties;
 pub mod read;
 pub mod remove;
-pub mod section_scanner;
 pub mod set;
-pub mod skills_lint;
 pub mod summary;
 pub mod tags;
 pub mod tasks;
