+++ crates/hyalo-mdlint/src/profiles/link.rs
