+++ crates/hyalo-mdlint/src/profiles/section_scanner.rs
